seperated auth middleware
