+++ server/routes/category.ts
@@ -7,7 +7,8 @@ import {
   updateCategory,
   removeCategory,
 } from '../controllers/category';
-import { isSignedin, isAuthenticated, isAdmin } from '../controllers/auth';
+import { isAdmin, isAuthenticated } from '../middlewares/authMiddleware';
+import { isSignedin } from '../controllers/auth';
 import { getUserById } from '../controllers/user';
 
 const router = express.Router();
